Add thirdparty handler
