+++ src/matrixroomhandler.ts
@@ -15,6 +15,7 @@ import * as log from "npmlog";
 import * as Bluebird from "bluebird";
 import { Util } from "./util";
 import { Provisioner } from "./provisioner";
+import { ThirdpartyHandler } from "./thirdpartyhandler";
 
 const ICON_URL = "https://matrix.org/_matrix/media/r0/download/matrix.org/mlxoESwIsTbJrfXyAAogrNxA";
 const HTTP_UNSUPPORTED = 501;
@@ -35,15 +36,16 @@ const JOIN_ROOM_SCHEDULE = [
 /* tslint:enable:no-magic-numbers */
 
 export class MatrixRoomHandler {
-
   private config: DiscordBridgeConfig;
   private bridge: Bridge;
   private discord: DiscordBot;
   private botUserId: string;
+  private thirdPartyHandler: ThirdpartyHandler;
   constructor (discord: DiscordBot, config: DiscordBridgeConfig, botUserId: string, private provisioner: Provisioner) {
     this.discord = discord;
     this.config = config;
     this.botUserId = botUserId;
+    this.thirdPartyHandler = this.discord.ThirdpartyHandler;
   }
 
   public get ThirdPartyLookup(): thirdPartyLookup {
@@ -352,7 +354,7 @@ export class MatrixRoomHandler {
 
   public tpGetLocation(protocol: string, fields: any): Promise<thirdPartyLocationResult[]> {
     log.info("MatrixRoomHandler", "Got location request ", protocol, fields);
-    const chans = this.discord.ThirdpartySearchForChannels(fields.guild_id, fields.channel_name);
+    const chans = this.thirdPartyHandler.SearchChannels(fields.guild_id, fields.channel_name);
     return Promise.resolve(chans);
   }
